#40 [0.2h] implement change requests
