+++ src/pages/ActivityLog.tsx
@@ -21,8 +21,8 @@ const ActivityLogTable: React.FC = () => {
 
   return (
     <div className="log-container">
-      <div className="log-export-wrapper">
-        <button className="ed-export-btn" onClick={() => logService.exportToCSV()}>
+      <div className="ed-card-header-flex">
+      <button className="ed-export-btn" onClick={() => logService.exportToCSV()}>
           CSV Export
         </button>
       </div>
