+++ src/services/csvService.ts
@@ -1,0 +1,31 @@
+export const csvService = {
+  /**
+   * @param headers 
+   * @param rows 
+   * @param filename
+   */
+  exportToCSV(headers: string[], rows: (string | number)[][], filename: string): void {
+    if (!rows || rows.length === 0) {
+      alert("Keine Daten zum Exportieren vorhanden.");
+      return;
+    }
+
+    const csvContent = [
+      headers.join(";"), 
+      ...rows.map(row => row.join(";"))
+    ].join("\n");
+    
+    const blob = new Blob(["\uFEFF" + csvContent], { type: 'text/csv;charset=utf-8;' });
+    const url = URL.createObjectURL(blob);
+    const link = document.createElement("a");
+    
+    link.href = url;
+    link.download = `${filename}.csv`;
+    
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+    
+    URL.revokeObjectURL(url);
+  }
+};
